+++ CHANGELOG.md
@@ -1,10 +1,16 @@
 # Change Log
 
+## 2.0.2
+- Updated to FormGenerator 2.1.0
+- Added more example fields
+- **Changed routes from `/projects` to `/formgenerator`**
+- Display the form data in an alert
+
 ## 2.0.1
-Fix issue with previous release.
+- Fix issue with previous release.
 
 ## 2.0.0
-Updated for UserFrosting v4.1.x
+- Updated for UserFrosting v4.1.x
 
 ## 1.0.0
-First release for UserFrosting v4.0.x
+- First release for UserFrosting v4.0.x